Vertex mapping ID is represented with a Long

+++ src/main/scala/gnormalizer/Vertex.scala
@@ -3,5 +3,5 @@ package gnormalizer
 object Vertex {
 
   type InputVertex = String
-  type VertexMapping = Int
+  type VertexMapping = Long
 }
